+++ django_airavata/apps/groups/static/django_airavata_groups/js/containers/GroupCreateContainer.vue
@@ -4,9 +4,8 @@
 
 <script>
 
-import GroupOwnerList from './GroupOwnerList.vue';
-import GroupMemberList from './GroupMemberList.vue';
-import GroupEditor from './GroupEditor.vue';
+import GroupList from '../group_components/GroupList.vue';
+import GroupEditor from '../group_components/GroupEditor.vue';
 
 import { models, services } from 'django-airavata-api'
 import { components as comps } from 'django-airavata-common-ui'
