+++ django_airavata/apps/groups/static/django_airavata_groups/js/group-create-entry-point.js
@@ -1,6 +1,6 @@
 import Vue from 'vue'
 import BootstrapVue from 'bootstrap-vue'
-import GroupCreateContainer from './groups_components/GroupCreateContainer.vue'
+import GroupCreateContainer from './containers/GroupCreateContainer.vue'
 // This is imported globally on the website so no need to include it again in this view
 // import 'bootstrap/dist/css/bootstrap.css'
 import 'bootstrap-vue/dist/bootstrap-vue.css'
