+++ django_airavata/apps/groups/static/django_airavata_groups/js/group-edit-entry-point.js
@@ -1,6 +1,6 @@
 import Vue from 'vue'
 import BootstrapVue from 'bootstrap-vue'
-import GroupEditContainer from './groups_components/GroupEditContainer.vue'
+import GroupEditContainer from './containers/GroupEditContainer.vue'
 // This is imported globally on the website so no need to include it again in this view
 // import 'bootstrap/dist/css/bootstrap.css'
 import 'bootstrap-vue/dist/bootstrap-vue.css'
