+++ django_airavata/apps/groups/static/django_airavata_groups/js/group-listing-entry-point.js
@@ -1,6 +1,6 @@
 import Vue from 'vue'
 import BootstrapVue from 'bootstrap-vue'
-import GroupsManageContainer from './groups_components/GroupsManageContainer.vue'
+import GroupsManageContainer from './containers/GroupsManageContainer.vue'
 // This is imported globally on the website so no need to include it again in this view
 // import 'bootstrap/dist/css/bootstrap.css'
 import 'bootstrap-vue/dist/bootstrap-vue.css'
@@ -9,10 +9,9 @@ Vue.use(BootstrapVue);
 
 new Vue({
   el: "#groups-owners-manage",
-  template: '<groups-manage-container v-bind:groupsDataOwners="groupsOwnersData" v-bind:groupsDataMembers="groupsMembersData"></groups-manage-container>',
+  template: '<groups-manage-container v-bind:groupsDataOwners="groupsOwnersData"></groups-manage-container>',
   data: {
       groupsOwnersData: null,
-      groupsMembersData: null,
   },
   components: {
       GroupsManageContainer
@@ -21,10 +20,6 @@ new Vue({
       if (this.$el.dataset.groupsOwnersData) {
           this.groupsOwnersData = JSON.parse(this.$el.dataset.groupsOwnersData);
           console.log(this.groupsOwnersData);
-      }
-      if(this.$el.dataset.groupsMembersData) {
-          this.groupsMembersData = JSON.parse(this.$el.dataset.groupsMembersData);
-          console.log(this.groupsMembersData);
       }
   }
 })
